requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
async-timeout==4.0.3
    # via redis
django==5.1.1
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements.in
pydantic==2.9.2
    # via -r requirements.in
pydantic-core==2.23.4
    # via pydantic
redis==5.0.8
    # via -r requirements.in
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via
    #   asgiref
    #   pydantic
    #   pydantic-core
